Alle JavaDoc Kommentare, kleinere Fixes

+++ DailyDataServer/src/main/java/com/pseandroid2/dailydataserver/onlineDatabase/ProjectParticipantService.java
@@ -31,6 +31,11 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * The Service for the ProjectParticipantsController which provides the logic.
+ * <p>
+ * It also generates the ids for projects.
+ */
 @Service
 public class ProjectParticipantService {
     private static final int MAX_PARTICIPANTS = 24;
@@ -38,6 +43,13 @@ public class ProjectParticipantService {
     private final ProjectRepository projectRepo;
     private long projectIDGenerator;
 
+    /**
+     * The Constructor for the ProjectParticipantsService.
+     *
+     * @param ppRepo      the ProjectParticipantsRepository, which provides operations to add and remove
+     *                    ProjectParticipants.
+     * @param projectRepo the projectRepository, to create, update and delete projects.
+     */
     public ProjectParticipantService(ProjectParticipantsRepository ppRepo, ProjectRepository projectRepo) {
         this.ppRepo = ppRepo;
         this.projectRepo = projectRepo;
@@ -45,6 +57,13 @@ public class ProjectParticipantService {
     }
 
 
+    /**
+     * a new project is going to be created. The user, who creates the project gets the admin.
+     *
+     * @param user        the user, who creates the project.
+     * @param projectInfo the initial for the project.
+     * @return the id for the new project.
+     */
     public long addProject(String user, String projectInfo) {
         Project project = new Project(projectIDGenerator, projectInfo);
         projectIDGenerator++;
@@ -55,6 +74,14 @@ public class ProjectParticipantService {
         return project.getProjectId();
     }
 
+    /**
+     * Adds a user to a project, if the project has less than MAX_PARTICIPANTS participants.
+     * If the user already participates the project, they also get the projectInitial.
+     *
+     * @param user      the user who wants to join a project.
+     * @param projectID the project, to which the user wants to join.
+     * @return an empty String, if the user can not join, the project initial, if the user joined the project.
+     */
     public String addUser(String user, long projectID) {
         if (ppRepo.existsById(new ProjectParticipantID(user, projectID))) {
             return projectRepo.findById(projectID).get().getProjectInfo();
@@ -66,14 +93,24 @@ public class ProjectParticipantService {
                 ppRepo.save(new ProjectParticipant(user, projectID, Role.PARTICIPANT, project.getParticipantJoin()));
 
                 project.setParticipantJoin(project.getParticipantJoin() + 1);
+                project.setLastUpdated(LocalDateTime.now());
                 projectRepo.save(project);
 
                 return project.getProjectInfo();
             }
-
         }
     }
 
+    /**
+     * a user leaves the project. If the admin leaves the project, the participant going to be the new admin, which
+     * participates for the longest time in the project.
+     * If the admin was the last person in the project, the project gets deleted.
+     * If the project still exists, it gets updated.≈
+     *
+     * @param user      the user, who wants to leave the project.
+     * @param projectId the project, which the user wants to leave.
+     * @return if the user could leave the project.
+     */
     public boolean leaveProject(String user, long projectId) {
         ProjectParticipant participant = ppRepo.findById(new ProjectParticipantID(user, projectId)).get();
         if (participant.getRole().equals(Role.ADMIN)) {
@@ -83,23 +120,42 @@ public class ProjectParticipantService {
             } else {
                 List<ProjectParticipant> projectParticipantList = ppRepo.findByProjectOrderByNumberOfJoinAsc(projectId);
                 projectParticipantList.get(1).setRole(Role.ADMIN);
+                ppRepo.save(projectParticipantList.get(1));
 
                 Project projectToUpdate = projectRepo.getById(projectId);
                 projectToUpdate.setLastUpdated(LocalDateTime.now());
                 projectRepo.save(projectToUpdate);
             }
+        } else {
+            ppRepo.deleteById(new ProjectParticipantID(user, projectId));
+            Project projectToUpdate = projectRepo.getById(projectId);
+            projectToUpdate.setLastUpdated(LocalDateTime.now());
+            projectRepo.save(projectToUpdate);
         }
-        ppRepo.deleteById(new ProjectParticipantID(user, projectId));
 
         return true;
     }
 
+    /**
+     * The admin can remove another user from their project.
+     * <p>
+     * If the user, which should be removed do not participate in the project, the method returns true as well.
+     * <p>
+     * This method checks, if the user, who wants to remove another user, is admin.
+     *
+     * @param user         the user, who wants to remove a user.
+     * @param projectId    the project where the user should be removed from.
+     * @param userToRemove the user, who is going to be removed.
+     * @return true, if the user could get removed, false, if the user can not get removed.
+     */
     public boolean removeOtherUser(String user, long projectId, String userToRemove) {
         ProjectParticipant participant = ppRepo.findById(new ProjectParticipantID(user, projectId)).get();
         if (!participant.getRole().equals(Role.ADMIN)) {
             return false;
         } else {
-            ppRepo.deleteById(new ProjectParticipantID(userToRemove, projectId));
+            if (ppRepo.existsById(new ProjectParticipantID(userToRemove, projectId))) {
+                ppRepo.deleteById(new ProjectParticipantID(userToRemove, projectId));
+            }
             Project projectToUpdate = projectRepo.getById(projectId);
             projectToUpdate.setLastUpdated(LocalDateTime.now());
             projectRepo.save(projectToUpdate);
@@ -107,6 +163,12 @@ public class ProjectParticipantService {
         }
     }
 
+    /**
+     * returns all users, who participate on a project. It also returns the admin.
+     *
+     * @param projectId the projectId, from which the user where recommended.
+     * @return all users, who participates in the project.
+     */
     public List<String> getParticipants(long projectId) {
         List<ProjectParticipant> participantsList = ppRepo.findByProject(projectId);
         List<String> participantName = new ArrayList<>();
@@ -116,6 +178,12 @@ public class ProjectParticipantService {
         return participantName;
     }
 
+    /**
+     * Provides the admin of a specified project.
+     *
+     * @param projectId the project, from which the admin is recommended.
+     * @return the admin.
+     */
     public String getAdmin(long projectId) {
         return ppRepo.findByProjectAndRoleIs(projectId, Role.ADMIN).getUser();
     }
